+++ packages/orchestrators/src/orchestrators/defs/triage_knowledge_queue/sensors.py
@@ -71,10 +71,6 @@ def poll_notion_for_triage(
         )
         publish_date_iso = publish_date_date.get("start") if publish_date_date else None
 
-        use_body_prop = row.get("properties", {}).get("Use page body", {})
-        use_body = bool(use_body_prop.get("checkbox", False))
-        raw_content_override = triage_notion.get_page_body_markdown(page_id) if use_body else ""
-
         last_edited = row.get("last_edited_time") or ""
         page_ids.append(page_id)
         run_requests.append(
@@ -92,7 +88,6 @@ def poll_notion_for_triage(
                             name=existing_name,
                             added_at_iso=added_at_iso,
                             publish_date_iso=publish_date_iso,
-                            raw_content_override=raw_content_override,
                         ),
                     }
                 ),